+++ src/core/io/types/uint64.cpp
@@ -10,28 +10,28 @@
 #include <stdexcept>
 #include <cmath>
 
-UInt64::UInt64(unsigned int _low, unsigned int _high) :
+UInt64::UInt64(uint _low, uint _high) :
         Binary64(_low, _high)
 {
 }
 
 UInt64 UInt64::fromNumber(double value)
 {
-    return UInt64(value, std::floor(value / 4294967296));
+    return UInt64(static_cast<uint>(value), static_cast<uint>(std::floor(value / 4294967296)));
 }
 
-UInt64 UInt64::parseUInt64(std::string str, unsigned int radix)
+UInt64 UInt64::parseUInt64(std::string str, uint radix)
 {
     throw std::logic_error("function not implemented");
     return UInt64();
 }
 
-void UInt64::setHigh(unsigned int high)
+void UInt64::setHigh(uint high)
 {
     internalHigh = high;
 }
 
-unsigned int UInt64::getHigh()
+uint UInt64::getHigh()
 {
     return internalHigh;
 }
@@ -41,7 +41,7 @@ double UInt64::toNumber()
     return (getHigh() * 4294967296) + low;
 }
 
-std::string UInt64::toString(unsigned int radix)
+std::string UInt64::toString(uint radix)
 {
     throw std::logic_error("function not implemented");
     return "";
